requirements.txt
python-dotenv==0.10.5
python-telegram-bot==11.1.0
redis==3.5.1
vk-api==11.7.0
